+++ sources/Struct.public_methods_definitions.c
@@ -196,7 +196,7 @@ int          oyStruct_CheckType      ( oyStruct_s        * obj,
       } else
         return 2;
 
-      if((intptr_t)inheritance < (intptr_t)oyOBJECT_MAX)
+      if((uint32_t)inheritance < (uint32_t)oyOBJECT_MAX)
       {
         WARNc1_S( "non plausible inheritance pointer: %s", 
                   oyStruct_GetInfo(obj,0) );
